modified

+++ Tetris_basic/GameConfig.cs
@@ -38,7 +38,7 @@ namespace Tetris_basic
 
         public int Xcoord { get; set; }
         public int Ycoord { get; set; }
-        public int Score { get; set; }
+        public long Score { get; set; }
         public int Level { get; set; }
 
         public GameConfig()
